backend/requirements.txt
fastapi==0.115.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.9
pymupdf==1.25.5
python-docx==1.1.0
pdfplumber
spacy==3.8.3
scikit-learn==1.4.2
sentence-transformers==3.0.0
torch==2.3.0
transformers==4.40.2
openai>=1.0.0